--- .github/workflows/template_change_test.yml
@@ -1,59 +1,0 @@
-# THIS ACTION IS FOR VALIDATION WITHIN TEMPLATE REPO, AND WILL BE
-# REMOVED UPON RUNNING RENAMETEMPLATE GRADLE TASK.
-
-name: Rename Template
-
-# This will cancel any in progress workflows for the same PR, if
-# multiple pushes happen in quick succession.
-concurrency:
-  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
-  cancel-in-progress: true
-
-on: pull_request
-
-jobs:
-  rename-template:
-    strategy:
-      matrix:
-        keepCustomizableDependencies: [true, false]
-      fail-fast: false
-
-    runs-on: macos-14
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Setup
-        uses: ./.github/actions/workflow_setup
-
-      - name: Set useHiltDependencies
-        run: |
-          sed -i '' 's/useHiltDependencies.* : true/useHiltDependencies : ${{ matrix.keepCustomizableDependencies }}/g' buildscripts/setup.gradle
-
-      - name: Set useRoomDependencies
-        run: |
-          sed -i '' 's/useRoomDependencies.* : true/useRoomDependencies : ${{ matrix.keepCustomizableDependencies }}/g' buildscripts/setup.gradle
-
-      - name: Set useRetrofitDependencies
-        run: |
-          sed -i '' 's/useRetrofitDependencies.* : true/useRetrofitDependencies : ${{ matrix.keepCustomizableDependencies }}/g' buildscripts/setup.gradle
-
-      - name: Set usePaparazziDependencies
-        run: |
-          sed -i '' 's/usePaparazziDependencies.* : true/usePaparazziDependencies : ${{ matrix.keepCustomizableDependencies }}/g' buildscripts/setup.gradle
-
-      - name: Set useRenovateDependencies
-        run: |
-          sed -i '' 's/useRenovateDependencies.* : true/useRenovateDependencies : ${{ matrix.keepCustomizableDependencies }}/g' buildscripts/setup.gradle
-
-      - name: Set useAndroidXR
-        run: |
-          sed -i '' 's/useAndroidXR.* : false/useAndroidXR : ${{ matrix.keepCustomizableDependencies }}/g' buildscripts/setup.gradle
-
-      - name: Rename
-        run: ./gradlew renameTemplate
-
-      - name: Lint Checks
-        run: ./gradlew detektAll lintKotlin lint
-
-      - name: Build
-        run: ./gradlew assembleDebug
