+++ app/src/main/java/com/juansanz/themoviedb/MainActivity.kt
@@ -1,4 +1,4 @@
-package template
+package com.juansanz.themoviedb
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -8,7 +8,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import template.theme.TemplateTheme
+import com.juansanz.themoviedb.theme.TheMovieDBTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -17,7 +17,7 @@ class MainActivity : ComponentActivity() {
         setContent {
             enableEdgeToEdge()
 
-            TemplateTheme {
+            TheMovieDBTheme {
                 Surface(
                     color = MaterialTheme.colorScheme.background,
                 ) {
