+++ app/src/main/java/com/juansanz/themoviedb/theme/Color.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MagicNumber")
 
-package template.theme
+package com.juansanz.themoviedb.theme
 
 import androidx.compose.ui.graphics.Color
