+++ app/src/main/java/com/juansanz/themoviedb/theme/Shape.kt
@@ -1,4 +1,4 @@
-package template.theme
+package com.juansanz.themoviedb.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Shapes
