+++ app/src/main/java/com/juansanz/themoviedb/theme/Type.kt
@@ -1,4 +1,4 @@
-package template.theme
+package com.juansanz.themoviedb.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
