+++ settings.gradle.kts
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "template"
+rootProject.name = "TheMovieDB"
 include(":app")
